# Performance improvements: Add Brainstorming Label to idea + broadcast lane update