--- a/lib/mindwendel/lanes.ex
+++ b/lib/mindwendel/lanes.ex
@@ -72,10 +72,7 @@ defmodule Mindwendel.Lanes do
   def get_lanes_for_brainstorming_with_labels_filtered(id) do
     {:ok, brainstorming} = Brainstormings.get_brainstorming(id)
 
-    filter_label =
-      if length(brainstorming.filter_labels_ids) > 0,
-        do: %{filter_labels_ids: brainstorming.filter_labels_ids},
-        else: %{}
+    filter_label = %{filter_labels_ids: brainstorming.filter_labels_ids}
 
     get_lanes_for_brainstorming(id, filter_label)
   end
@@ -89,7 +86,7 @@ defmodule Mindwendel.Lanes do
       [%Lane{}, ...]
 
   """
-  def get_lanes_for_brainstorming(id, filters \\ %{}) do
+  def get_lanes_for_brainstorming(id, filters \\ %{filter_labels_ids: []}) do
     lane_query =
       from lane in Lane,
         where: lane.brainstorming_id == ^id,
@@ -105,6 +102,10 @@ defmodule Mindwendel.Lanes do
     |> Repo.preload(ideas: {ideas_advanced_query, [:link, :likes, :idea_labels, :files]})
   end
 
+  defp build_ideas_query_with_filter(%{filter_labels_ids: []}) do
+    from(idea in Idea)
+  end
+
   defp build_ideas_query_with_filter(%{filter_labels_ids: filter_labels_ids}) do
     distinct_ideas =
       from idea in Idea,
@@ -116,10 +117,6 @@ defmodule Mindwendel.Lanes do
     from(i in subquery(distinct_ideas),
       order_by: [asc: i.position_order]
     )
-  end
-
-  defp build_ideas_query_with_filter(%{} = _filters) do
-    from(idea in Idea)
   end
 
   @doc """