--- a/lib/mindwendel_web/live/idea_live/card_component.html.heex
+++ b/lib/mindwendel_web/live/idea_live/card_component.html.heex
@@ -97,7 +97,6 @@
             data-testid={brainstorming_idea_label.id}
             phx-click="add_idea_label_to_idea"
             phx-target={@myself}
-            phx-value-idea-id={@idea.id}
             title={"Label #{brainstorming_idea_label.name}"}
             phx-value-idea-label-id={brainstorming_idea_label.id}
           >
@@ -116,7 +115,6 @@
             data-testid={brainstorming_idea_label.id}
             phx-click="remove_idea_label_from_idea"
             phx-target={@myself}
-            phx-value-idea-id={@idea.id}
             title={"Label #{brainstorming_idea_label.name}"}
             phx-value-idea-label-id={brainstorming_idea_label.id}
           >