--- a/lib/mindwendel_web/live/idea_live/card_component.html.heex
+++ b/lib/mindwendel_web/live/idea_live/card_component.html.heex
@@ -13,7 +13,6 @@
         class="float-end ms-3 mb-3"
         phx-click="delete_idea"
         phx-target={@myself}
-        phx-value-id={@idea.id}
         title={gettext("Delete idea")}
         data-confirm={gettext("Are you sure you want to delete this idea?")}
       >
@@ -143,11 +142,11 @@
       </span>
       <span class="me-1 text-dark">{length(@idea.likes)}</span>
       <%= if Mindwendel.Likes.exists_user_in_likes?(@idea.likes, @current_user.id) do %>
-        <.link phx-click="unlike" phx-target={@myself} phx-value-id={@idea.id} title="Unlike">
+        <.link phx-click="unlike" phx-target={@myself} title="Unlike">
           <i class="bi-star-fill"></i>
         </.link>
       <% else %>
-        <.link phx-click="like" phx-target={@myself} phx-value-id={@idea.id} title="Like">
+        <.link phx-click="like" phx-target={@myself} title="Like">
           <i class="bi-star"></i>
         </.link>
       <% end %>