--- a/test/mindwendel/brainstormings_test.exs
+++ b/test/mindwendel/brainstormings_test.exs
@@ -322,7 +322,7 @@ defmodule Mindwendel.BrainstormingsTest do
 
       like = Factory.insert!(:like, idea: idea)
 
-      {:ok, idea} =
+      {:ok, _idea_idea_label} =
         IdeaLabels.add_idea_label_to_idea(idea, Enum.at(brainstorming.labels, 0))
 
       idea = idea |> Repo.preload([:idea_labels])