--- a/test/mindwendel/brainstormings_test.exs
+++ b/test/mindwendel/brainstormings_test.exs
@@ -323,7 +323,7 @@ defmodule Mindwendel.BrainstormingsTest do
       like = Factory.insert!(:like, idea: idea)
 
       {:ok, _idea_idea_label} =
-        IdeaLabels.add_idea_label_to_idea(idea, Enum.at(brainstorming.labels, 0))
+        IdeaLabels.add_idea_label_to_idea(idea, Enum.at(brainstorming.labels, 0).id)
 
       idea = idea |> Repo.preload([:idea_labels])