--- a/test/mindwendel/idea_labels_test.exs
+++ b/test/mindwendel/idea_labels_test.exs
@@ -9,7 +9,7 @@ defmodule Mindwendel.IdeaLabelsTest do
 
   setup do
     brainstorming = Factory.insert!(:brainstorming, %{labels: [Factory.build(:idea_label)]})
-    idea_label = brainstorming.labels |> Enum.at(0)
+    [idea_label] = brainstorming.labels
     lane = Enum.at(brainstorming.lanes, 0)
     idea = Factory.insert!(:idea, %{brainstorming: brainstorming, lane: lane})
 
@@ -18,16 +18,16 @@ defmodule Mindwendel.IdeaLabelsTest do
 
   describe "#add_idea_label_to_idea" do
     test "adds IdeaLabel to Idea", %{idea_label: idea_label, idea: idea} do
-      {:ok, idea_changed} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
+      {:ok, _idea_idea_label} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
 
-      assert idea_changed.idea_labels |> Enum.count() == 1
-      assert Repo.all(IdeaIdeaLabel) |> Enum.count() == 1
+      assert [idea_label] == labels_of(idea)
+      assert Repo.count(IdeaIdeaLabel) == 1
     end
 
     test "creates one IdeaIdeaLabel", %{idea_label: idea_label, idea: idea} do
-      {:ok, _idea_changed} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
+      {:ok, _idea_idea_label} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
 
-      assert Repo.all(IdeaIdeaLabel) |> Enum.count() == 1
+      assert Repo.count(IdeaIdeaLabel) == 1
 
       idea_idea_label = Repo.one(IdeaIdeaLabel)
       assert idea_idea_label.idea_label_id == idea_label.id
@@ -35,26 +35,29 @@ defmodule Mindwendel.IdeaLabelsTest do
     end
 
     test "does not create additional IdeaLabel", %{idea_label: idea_label, idea: idea} do
-      assert Repo.all(IdeaLabel) |> Enum.count() == 1
+      assert Repo.count(IdeaLabel) == 1
 
-      {:ok, _idea_changed} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
+      {:ok, _idea_idea_label} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
 
-      assert Repo.all(IdeaLabel) |> Enum.count() == 1
+      assert Repo.count(IdeaLabel) == 1
       assert Repo.one(IdeaLabel) == idea_label
     end
 
     @tag :skip
     test "does not add the same IdeaLabel twice to Idea", %{idea_label: idea_label, idea: idea} do
       # Calling this method twice does not fail and does not create duplicates
-      {:ok, idea_after_method_call_1} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
-      {:ok, idea_after_method_call_2} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
+      {:ok, idea_idea_label_after_method_call_1} =
+        IdeaLabels.add_idea_label_to_idea(idea, idea_label)
+
+      {:ok, idea_idea_label_after_method_call_2} =
+        IdeaLabels.add_idea_label_to_idea(idea, idea_label)
 
       # There should still be only one IdeaIdeaLabel
-      assert Repo.all(IdeaIdeaLabel) |> Enum.count() == 1
+      assert Repo.count(IdeaIdeaLabel) == 1
 
-      assert Repo.all(IdeaLabel) |> Enum.count() == 1
+      assert Repo.count(IdeaLabel) == 1
 
-      assert idea_after_method_call_1 == idea_after_method_call_2
+      assert idea_idea_label_after_method_call_1 == idea_idea_label_after_method_call_2
     end
 
     @tag :skip
@@ -88,16 +91,16 @@ defmodule Mindwendel.IdeaLabelsTest do
 
     @tag :skip
     test "does not create idea_labels without brainstorming", %{idea: idea} do
-      assert Repo.all(IdeaIdeaLabel) |> Enum.count() == 0
+      assert Repo.count(IdeaIdeaLabel) == 0
       idea = Repo.preload(idea, [:idea_labels, :idea_idea_labels])
       idea_label = idea.brainstorming.labels |> Enum.at(0)
 
       IdeaLabels.add_idea_label_to_idea(idea, idea_label)
 
-      assert Repo.all(IdeaIdeaLabel) |> Enum.count() == 1
+      assert Repo.count(IdeaIdeaLabel) == 1
       assert Repo.one(IdeaIdeaLabel).idea_id == idea.id
       assert Repo.one(IdeaIdeaLabel).idea_label_id == idea_label.id
-      assert Repo.all(IdeaLabel) |> Enum.count() == 5
+      assert Repo.count(IdeaLabel) == 5
 
       assert Enum.empty?(
                Repo.all(
@@ -110,8 +113,9 @@ defmodule Mindwendel.IdeaLabelsTest do
 
   describe "#delete_idea_label_from_idea" do
     setup %{idea_label: idea_label, idea: idea} do
-      {:ok, idea} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
-      assert Repo.all(IdeaIdeaLabel) |> Enum.count() == 1
+      {:ok, _idea_idea_label} = IdeaLabels.add_idea_label_to_idea(idea, idea_label)
+      idea = idea |> Repo.reload!() |> Repo.preload(:idea_labels)
+      assert Repo.count(IdeaIdeaLabel) == 1
       %{idea: idea}
     end
 
@@ -130,5 +134,13 @@ defmodule Mindwendel.IdeaLabelsTest do
 
       assert Enum.empty?(Repo.all(IdeaIdeaLabel))
     end
+  end
+
+  defp labels_of(idea_record) do
+    Repo.all(
+      from idea_label in IdeaLabel,
+        join: idea in assoc(idea_label, :ideas),
+        where: idea.id == ^idea_record.id
+    )
   end
 end