This was mostly me looking at SQL emitted when adding a new label.

This PR reduces the total amount of SQL queries issued for adding a label **from 20 to 14**
<details>
<summary>Log for add label before optimization</summary>

```
[debug] HANDLE EVENT "add_idea_label_to_idea" in MindwendelWeb.BrainstormingLive.Show
  Component: MindwendelWeb.IdeaLive.CardComponent
  Parameters: %{"idea-id" => "7c9dca19-94df-4653-b2db-1c2cc49edacc", "idea-label-id" => "84aedeef-eb27-4cb4-8804-05eca963b6e0"}
[debug] QUERY OK source="ideas" db=0.3ms idle=484.8ms
SELECT i0."id", i0."body", i0."position_order", i0."username", i0."comments_count", i0."user_id", i0."brainstorming_id", i0."lane_id", i0."inserted_at", i0."updated_at" FROM "ideas" AS i0 WHERE (i0."id" = $1) ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="idea_files" db=0.2ms idle=485.4ms
SELECT i0."id", i0."name", i0."path", i0."file_type", i0."idea_id", i0."inserted_at", i0."updated_at", i0."idea_id" FROM "idea_files" AS i0 WHERE (i0."idea_id" = $1) ORDER BY i0."idea_id" ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="idea_comments" db=0.2ms idle=485.5ms
SELECT i0."id", i0."idea_id", i0."user_id", i0."body", i0."username", i0."inserted_at", i0."updated_at", i0."idea_id" FROM "idea_comments" AS i0 WHERE (i0."idea_id" = $1) ORDER BY i0."idea_id", i0."inserted_at" DESC ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="links" db=0.3ms idle=485.6ms
SELECT l0."id", l0."url", l0."title", l0."description", l0."img_preview_url", l0."idea_id", l0."inserted_at", l0."updated_at", l0."idea_id" FROM "links" AS l0 WHERE (l0."idea_id" = $1) ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="idea_labels" db=1.1ms idle=485.4ms
SELECT i0."id", i0."name", i0."color", i0."position_order", i0."brainstorming_id", i0."inserted_at", i0."updated_at", i1."idea_id"::uuid FROM "idea_labels" AS i0 INNER JOIN "idea_idea_labels" AS i1 ON i0."id" = i1."idea_label_id" WHERE (i1."idea_id" = ANY($1)) ORDER BY i1."idea_id"::uuid [["7c9dca19-94df-4653-b2db-1c2cc49edacc"]]
[debug] QUERY OK source="idea_labels" db=0.4ms idle=486.7ms
SELECT i0."id", i0."name", i0."color", i0."position_order", i0."brainstorming_id", i0."inserted_at", i0."updated_at" FROM "idea_labels" AS i0 WHERE (i0."id" = $1) ["84aedeef-eb27-4cb4-8804-05eca963b6e0"]
[debug] QUERY OK db=0.1ms idle=487.3ms
begin []
[debug] QUERY OK source="idea_idea_labels" db=0.4ms
INSERT INTO "idea_idea_labels" ("idea_label_id","idea_id","inserted_at","updated_at") VALUES ($1,$2,$3,$4) ["84aedeef-eb27-4cb4-8804-05eca963b6e0", "7c9dca19-94df-4653-b2db-1c2cc49edacc", ~N[2024-11-30 18:37:32], ~N[2024-11-30 18:37:32]]
[debug] QUERY OK db=7.6ms
commit []
[debug] QUERY OK source="brainstormings" db=0.1ms idle=495.8ms
SELECT b0."id", b0."name", b0."option_allow_manual_ordering", b0."admin_url_id", b0."last_accessed_at", b0."filter_labels_ids", b0."creating_user_id", b0."inserted_at", b0."updated_at" FROM "brainstormings" AS b0 WHERE (b0."id" = $1) ["25a35202-c68d-4278-af69-382849134847"]
[debug] QUERY OK source="idea_labels" db=0.2ms idle=11.1ms
SELECT i0."id", i0."name", i0."color", i0."position_order", i0."brainstorming_id", i0."inserted_at", i0."updated_at", i0."brainstorming_id" FROM "idea_labels" AS i0 WHERE (i0."brainstorming_id" = $1) ORDER BY i0."brainstorming_id", i0."position_order" ["25a35202-c68d-4278-af69-382849134847"]
[debug] QUERY OK source="users" db=0.4ms idle=496.1ms
SELECT u0."id", u0."username", u0."inserted_at", u0."updated_at", b1."brainstorming_id"::uuid FROM "users" AS u0 INNER JOIN "brainstorming_moderating_users" AS b1 ON u0."id" = b1."user_id" WHERE (b1."brainstorming_id" = ANY($1)) ORDER BY b1."brainstorming_id"::uuid [["25a35202-c68d-4278-af69-382849134847"]]
[debug] QUERY OK source="users" db=0.4ms idle=373.5ms
SELECT u0."id", u0."username", u0."inserted_at", u0."updated_at", b1."brainstorming_id"::uuid FROM "users" AS u0 INNER JOIN "brainstorming_users" AS b1 ON u0."id" = b1."user_id" WHERE (b1."brainstorming_id" = ANY($1)) ORDER BY b1."brainstorming_id"::uuid [["25a35202-c68d-4278-af69-382849134847"]]
[debug] QUERY OK source="brainstormings" db=1.0ms idle=11.1ms
UPDATE "brainstormings" SET "last_accessed_at" = $1, "updated_at" = $2 WHERE "id" = $3 [~U[2024-11-30 18:37:32Z], ~N[2024-11-30 18:37:32], "25a35202-c68d-4278-af69-382849134847"]
[debug] QUERY OK source="lanes" db=0.2ms idle=12.3ms
SELECT l0."id", l0."name", l0."position_order", l0."brainstorming_id", l0."inserted_at", l0."updated_at" FROM "lanes" AS l0 WHERE (l0."brainstorming_id" = $1) ORDER BY l0."position_order", l0."inserted_at" ["25a35202-c68d-4278-af69-382849134847"]
[debug] QUERY OK source="ideas" db=0.2ms idle=12.6ms
SELECT i0."id", i0."body", i0."position_order", i0."username", i0."comments_count", i0."user_id", i0."brainstorming_id", i0."lane_id", i0."inserted_at", i0."updated_at", i0."lane_id" FROM "ideas" AS i0 WHERE (i0."lane_id" = ANY($1)) ORDER BY i0."lane_id", i0."position_order", i0."inserted_at" [["ef8dbd09-f1be-410b-8e35-298ffaaf1a09", "69f02828-b1f7-4b73-a8a7-6177090fda22"]]
[debug] QUERY OK source="idea_files" db=0.1ms idle=3.3ms
SELECT i0."id", i0."name", i0."path", i0."file_type", i0."idea_id", i0."inserted_at", i0."updated_at", i0."idea_id" FROM "idea_files" AS i0 WHERE (i0."idea_id" = ANY($1)) ORDER BY i0."idea_id" [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] QUERY OK source="links" db=0.1ms idle=3.1ms
SELECT l0."id", l0."url", l0."title", l0."description", l0."img_preview_url", l0."idea_id", l0."inserted_at", l0."updated_at", l0."idea_id" FROM "links" AS l0 WHERE (l0."idea_id" = ANY($1)) [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] QUERY OK source="idea_labels" db=0.4ms idle=12.4ms
SELECT i0."id", i0."name", i0."color", i0."position_order", i0."brainstorming_id", i0."inserted_at", i0."updated_at", i1."idea_id"::uuid FROM "idea_labels" AS i0 INNER JOIN "idea_idea_labels" AS i1 ON i0."id" = i1."idea_label_id" WHERE (i1."idea_id" = ANY($1)) ORDER BY i1."idea_id"::uuid [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] QUERY OK source="likes" db=0.4ms idle=11.9ms
SELECT l0."id", l0."idea_id", l0."user_id", l0."inserted_at", l0."updated_at", l0."idea_id" FROM "likes" AS l0 WHERE (l0."idea_id" = ANY($1)) ORDER BY l0."idea_id" [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] Replied in 15ms
```

</details>
<details>
<summary>Log for add label after optimization</summary>

```
[debug] HANDLE EVENT "add_idea_label_to_idea" in MindwendelWeb.BrainstormingLive.Show
  Component: MindwendelWeb.IdeaLive.CardComponent
  Parameters: %{"idea-id" => "7c9dca19-94df-4653-b2db-1c2cc49edacc", "idea-label-id" => "84aedeef-eb27-4cb4-8804-05eca963b6e0"}
[debug] QUERY OK source="ideas" db=0.5ms idle=1617.1ms
SELECT i0."id", i0."body", i0."position_order", i0."username", i0."comments_count", i0."user_id", i0."brainstorming_id", i0."lane_id", i0."inserted_at", i0."updated_at" FROM "ideas" AS i0 WHERE (i0."id" = $1) ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="idea_labels" db=0.3ms idle=624.0ms
SELECT i0."id", i0."name", i0."color", i0."position_order", i0."brainstorming_id", i0."inserted_at", i0."updated_at", i1."idea_id"::uuid FROM "idea_labels" AS i0 INNER JOIN "idea_idea_labels" AS i1 ON i0."id" = i1."idea_label_id" WHERE (i1."idea_id" = ANY($1)) ORDER BY i1."idea_id"::uuid [["7c9dca19-94df-4653-b2db-1c2cc49edacc"]]
[debug] QUERY OK source="idea_files" db=0.2ms idle=617.9ms
SELECT i0."id", i0."name", i0."path", i0."file_type", i0."idea_id", i0."inserted_at", i0."updated_at", i0."idea_id" FROM "idea_files" AS i0 WHERE (i0."idea_id" = $1) ORDER BY i0."idea_id" ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="idea_comments" db=0.3ms idle=618.0ms
SELECT i0."id", i0."idea_id", i0."user_id", i0."body", i0."username", i0."inserted_at", i0."updated_at", i0."idea_id" FROM "idea_comments" AS i0 WHERE (i0."idea_id" = $1) ORDER BY i0."idea_id", i0."inserted_at" DESC ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="links" db=0.2ms idle=618.1ms
SELECT l0."id", l0."url", l0."title", l0."description", l0."img_preview_url", l0."idea_id", l0."inserted_at", l0."updated_at", l0."idea_id" FROM "links" AS l0 WHERE (l0."idea_id" = $1) ["7c9dca19-94df-4653-b2db-1c2cc49edacc"]
[debug] QUERY OK source="idea_labels" db=0.2ms idle=618.7ms
SELECT i0."id", i0."name", i0."color", i0."position_order", i0."brainstorming_id", i0."inserted_at", i0."updated_at" FROM "idea_labels" AS i0 WHERE (i0."id" = $1) ["84aedeef-eb27-4cb4-8804-05eca963b6e0"]
[debug] QUERY OK source="idea_idea_labels" db=7.9ms queue=0.1ms idle=619.1ms
INSERT INTO "idea_idea_labels" ("idea_label_id","idea_id","inserted_at","updated_at") VALUES ($1,$2,$3,$4) ["84aedeef-eb27-4cb4-8804-05eca963b6e0", "7c9dca19-94df-4653-b2db-1c2cc49edacc", ~N[2024-11-30 19:37:49], ~N[2024-11-30 19:37:49]]
[debug] QUERY OK source="brainstormings" db=0.1ms idle=627.4ms
SELECT b0."id", b0."name", b0."option_allow_manual_ordering", b0."admin_url_id", b0."last_accessed_at", b0."filter_labels_ids", b0."creating_user_id", b0."inserted_at", b0."updated_at" FROM "brainstormings" AS b0 WHERE (b0."id" = $1) ["25a35202-c68d-4278-af69-382849134847"]
[debug] QUERY OK source="lanes" db=0.1ms idle=627.7ms
SELECT l0."id", l0."name", l0."position_order", l0."brainstorming_id", l0."inserted_at", l0."updated_at" FROM "lanes" AS l0 WHERE (l0."brainstorming_id" = $1) ORDER BY l0."position_order", l0."inserted_at" ["25a35202-c68d-4278-af69-382849134847"]
[debug] QUERY OK source="ideas" db=0.2ms idle=628.0ms
SELECT i0."id", i0."body", i0."position_order", i0."username", i0."comments_count", i0."user_id", i0."brainstorming_id", i0."lane_id", i0."inserted_at", i0."updated_at", i0."lane_id" FROM "ideas" AS i0 WHERE (i0."lane_id" = ANY($1)) ORDER BY i0."lane_id", i0."position_order", i0."inserted_at" [["ef8dbd09-f1be-410b-8e35-298ffaaf1a09", "69f02828-b1f7-4b73-a8a7-6177090fda22"]]
[debug] QUERY OK source="idea_labels" db=0.3ms idle=11.1ms
SELECT i0."id", i0."name", i0."color", i0."position_order", i0."brainstorming_id", i0."inserted_at", i0."updated_at", i1."idea_id"::uuid FROM "idea_labels" AS i0 INNER JOIN "idea_idea_labels" AS i1 ON i0."id" = i1."idea_label_id" WHERE (i1."idea_id" = ANY($1)) ORDER BY i1."idea_id"::uuid [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] QUERY OK source="likes" db=0.3ms idle=10.5ms
SELECT l0."id", l0."idea_id", l0."user_id", l0."inserted_at", l0."updated_at", l0."idea_id" FROM "likes" AS l0 WHERE (l0."idea_id" = ANY($1)) ORDER BY l0."idea_id" [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] QUERY OK source="idea_files" db=0.2ms idle=10.7ms
SELECT i0."id", i0."name", i0."path", i0."file_type", i0."idea_id", i0."inserted_at", i0."updated_at", i0."idea_id" FROM "idea_files" AS i0 WHERE (i0."idea_id" = ANY($1)) ORDER BY i0."idea_id" [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] QUERY OK source="links" db=0.2ms idle=10.8ms
SELECT l0."id", l0."url", l0."title", l0."description", l0."img_preview_url", l0."idea_id", l0."inserted_at", l0."updated_at", l0."idea_id" FROM "links" AS l0 WHERE (l0."idea_id" = ANY($1)) [["2dfd9635-f0e0-44e3-8acf-f080c681b07f", "f8b75443-8214-465a-af87-f0f87f1d3512", "30ca05a6-71e0-4fe1-9acf-9a99beaf50f8", "7c9dca19-94df-4653-b2db-1c2cc49edacc", "7bff94f2-e3e3-440d-8346-4c68a2ea00ab"]]
[debug] Replied in 12ms
```

</details>
That number is still relatively high (imo) but this is due to the decision of broadcasting a full "Lane change" instead of the incremental equivalent update (which I didn't wanna touch here/change without discussion).

The PR mainly does 2 things:
* limit what we do to add the label, as we discard the `idea` struct afterwards anyhow
* on the broadcast event, limit what brainstorming loads as all its preloads don't matter

There is some further potential here, I'll highlight some and will check out some more one of these days for similar patterns.