@PragTob 2024-11-30
I think we can easily remove 2 more database queries:

```elixir
  def handle_event(
        "add_idea_label_to_idea",
        %{
          "idea-id" => idea_id,
          "idea-label-id" => idea_label_id
        },
        socket
      ) do
    idea = Ideas.get_idea!(idea_id)
    idea_label = IdeaLabels.get_idea_label(idea_label_id)

    IdeaLabels.add_idea_label_to_idea(idea, idea_label)
    {:noreply, socket}
  end
```

We don't need to load the idea or idea_label - we only need the ids. However, later `IdeaLabels.add_idea_label_to_idea` needs the brainstorming_id of the `idea` to broadcast the changes. But I guess it's in the socket as well (I haven't checked yet) so we could pass that on explicitly and remove the 2 queries here as well. That'd bring the query count down to 12 :)