@JannikStreek 2024-11-30
Thanks, thats a good start of cleaning up! All the broadcasts can be improved a lot, we often load e.g. an idea or lane but in the end we just need an id. I think that happens actually in more places.

> but this is due to the decision of broadcasting a full "Lane change" instead of the incremental equivalent update
This is an interesting one, where I am open for feedback. The reason why we often broadcast a whole lane or all lanes is, that if you drag an idea, you change the complete ordering of all ideas. So it's not enough to just update idea currently. This could be done just for content changes. But if the position is changed, we need to update at least the lane where the idea is on and maybe even more lanes, if the idea was moved from lane A to B. Thats what makes it complicated and this is a very easy but improvable solution.

One solution here would be to do the ordering on the socket / client side and not to rely on ordering them before the broadcast happens. Then I guess (?) just updating one idea would be sufficient. Feely free to have a look at this.