@JannikStreek 2024-12-19
`Lanes.get_lanes_for_brainstorming_with_labels_filtered` also worth to look at as its required an id but often the brainstorming is already loaded before, so we are doing the same query 2 times.