@JannikStreek 2024-12-19
Also interesting question in general: Passing ids vs passing objects (refs). Passing the id, you always need to refetch the object in the designated method. On the other hand, always passing the full object might lead to strange errors as for an update method you need the most recent object usually (just happened to me).