@PragTob 2024-12-23
Yeah it's all a big game of tradeoffs.

* **lanes and position updates** - the emitted event could just be "this card moved from this position to this position" and each client just does a quick manipulation of their in memory (or client side) data structure. It's a case that often happens and should be doable. It should be less load on the machines and also far less data to broadcast.
However:
  * You always have the danger of things running out of sync with incremental errors - fully reloading gets rid of that
  * it's a special implementation every time around, if the performance isn't an issue, which I'm not sure it is right now - all the effort and potential bugs maybe/probably aren't worth it - updating everything everywhere is easier

In short, complexy and increased chances for bugs vs. lower load on the system and quicker response times. For the kind of project this is, I'd assume that going with the simpler "safer" implementation is fine. That said, it feels a bit "wasteful" in my mind.

* **ids vs. objects** - it's an interesting/case by case basis. My default is "do I need the whole object? No? Then let's pass an id?" that said, passing the object is a safe default/assumption and only switching to the ID when needed. Here I just noticed there were a bunch of queries and so I went searching/hunting :)
Passing the object can also be preferable since you avoid double fetching (i.e. 2 funtions get passed the id, but load their own record) and you can do a preload data for all needed associations (that said, if an association is already present it won't be refetched by default by `Repo.preload`).

Some functions are also built to handle both through a pattern match in the function head to deal with both being passed an id or being passed the full record.